move grapher scorer
