+++ packages/perseus/src/widgets/grapher/grapher.tsx
@@ -3,6 +3,12 @@ import {
     vector as kvector,
     point as kpoint,
 } from "@khanacademy/kmath";
+import {GrapherUtil} from "@khanacademy/perseus-core";
+import {
+    scoreGrapher,
+    type PerseusGrapherRubric,
+    type PerseusGrapherUserInput,
+} from "@khanacademy/perseus-score";
 import * as React from "react";
 import _ from "underscore";
 
@@ -20,14 +26,13 @@ import {getInteractiveBoxFromSizeClass} from "../../util/sizing-utils";
 /* Mixins. */
 import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/grapher/grapher-ai-utils";
 
-import scoreGrapher from "./score-grapher";
 import {
     DEFAULT_GRAPHER_PROPS,
     chooseType,
     defaultPlotProps,
-    functionForType,
     getGridAndSnapSteps,
     maybePointsFromNormalized,
+    movableToGraphieMovable,
     typeToButton,
 } from "./util";
 
@@ -40,10 +45,6 @@ import type {
     MarkingsType,
     PerseusGrapherWidgetOptions,
 } from "@khanacademy/perseus-core";
-import type {
-    PerseusGrapherRubric,
-    PerseusGrapherUserInput,
-} from "@khanacademy/perseus-score";
 import type {PropsFor} from "@khanacademy/wonder-blocks-core";
 
 // @ts-expect-error - TS2339 - Property 'MovablePoint' does not exist on type 'typeof Graphie'.
@@ -137,8 +138,9 @@ class FunctionGrapher extends React.Component<FunctionGrapherProps> {
         }
 
         const functionProps = model.getPropsForCoeffs(coeffs, xRange);
+        const Movable = movableToGraphieMovable[model.movable];
         return (
-            <model.Movable
+            <Movable
                 {...functionProps}
                 key={this.props.model.url}
                 range={xRange}
@@ -595,7 +597,7 @@ class Grapher extends React.Component<Props> implements Widget {
                 setup: this._setupGraphie,
             },
             onChange: this.handlePlotChanges,
-            model: type && functionForType(type),
+            model: type && GrapherUtil.functionForType(type),
             coords: coords,
             asymptote: asymptote,
             static: this.props.static,
